training-session-and-assessment-service/requirements.txt
fastapi==0.90.1
pre-commit==3.0.4
psycopg2==2.9.5
python-dotenv==0.21.1
SQLAlchemy==2.0.3
uvicorn==0.20.0
